+++ config/cache.config.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Common;
 
-use Doctrine\Common\Cache\Cache as DoctrineCache;
 use Predis\Client as PredisClient;
 use Predis\ClientInterface as PredisClientInterface;
 use Psr\Cache\CacheItemPoolInterface as PsrCache;
@@ -14,7 +13,6 @@ return [
     'dependencies' => [
         'factories' => [
             PsrCache::class => Cache\CacheFactory::class,
-            DoctrineCache::class => Cache\DoctrineCacheFactory::class,
             PredisClientInterface::class => Cache\RedisFactory::class,
         ],
         'aliases' => [